fix: update boilerplate with updated plugin configuration
