--- .eslintrc
@@ -1,20 +1,0 @@
-{
-    "env": {
-        "node": true,
-        "es6": true,
-        "jest": true
-    },
-    "parserOptions": {
-        "ecmaVersion": 2018
-    },
-    "plugins": [
-        "jest"
-    ],
-    "extends": [
-        "plugin:jest/recommended",
-        "eslint:recommended"
-    ],
-    "rules": {
-        "no-var": 2
-    }
-}
