+++ .eslintrc.json
@@ -1,0 +1,8 @@
+{
+    "extends": "@adobe/eslint-config-aio-lib-config",
+    "settings": {
+        "jsdoc": {
+          "ignorePrivate": true
+        }
+    }
+}
