+++ .npmrc
@@ -1,0 +1,2 @@
+package-lock=true
+tag-version-prefix=""
