+++ jest.setup.js
@@ -10,12 +10,10 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const { stdout } = require('stdout-stderr')
+const { stdout, stderr } = require('stdout-stderr')
 
 jest.setTimeout(30000)
 
-beforeEach(() => {
-  stdout.start()
-  jest.clearAllMocks()
-})
-afterEach(() => { stdout.stop() })
+// trap console log
+beforeEach(() => { stdout.start(); stderr.start() })
+afterEach(() => { stdout.stop(); stderr.stop() })
