+++ package.json
@@ -1,29 +1,30 @@
 {
-  "name": "@adobe/PLUGINNAME",
+  "name": "@adobe/aio-cli-plugin-PLUGINNAME",
   "version": "0.0.0",
   "dependencies": {
-    "@adobe/aio-cna-core-config": "^1.0.14",
-    "@oclif/command": "^1",
-    "@oclif/config": "^1",
-    "@oclif/errors": "^1.1.2",
-    "debug": "^4.1.0"
+    "@adobe/aio-lib-core-config": "^1.2.8",
+    "@adobe/aio-lib-core-logging": "^1.1.0",
+    "@oclif/command": "^1.6.1",
+    "@oclif/config": "^1.15.1",
+    "@oclif/errors": "^1.1.2"
   },
   "devDependencies": {
-    "@oclif/dev-cli": "^1.21.3",
-    "@oclif/plugin-help": "^2.1.4",
-    "@oclif/test": "^1",
-    "acorn": "^6.1.1",
-    "chalk": "^2.4.1",
-    "codecov": "^3.2.0",
-    "eslint": "^6.0.1",
+    "@adobe/eslint-config-aio-lib-config": "^1.2.0",
+    "@oclif/dev-cli": "^1",
+    "@oclif/plugin-help": "^2",
+    "acorn": "^7",
+    "chalk": "^4",
+    "eslint": "^6.2.2",
     "eslint-config-oclif": "^3.1.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-jest": "22.9.0",
-    "eslint-plugin-node": "^9.0.0",
-    "eslint-plugin-promise": "^4.0.0",
-    "eslint-plugin-standard": "^4.0.0",
-    "jest": "^24.1.0",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-jest": "^23.13.2",
+    "eslint-plugin-jsdoc": "^21.0.0",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
+    "execa": "^4",
+    "jest": "^24",
     "jest-haste-map": "^24.5.0",
     "jest-junit": "^6.0.0",
     "jest-resolve": "^24.5.0",
@@ -42,14 +43,15 @@
   "license": "Apache-2.0",
   "oclif": {
     "commands": "./src/commands",
-    "bin": "./bin/run",
+    "bin": "aio",
     "devPlugins": [
       "@oclif/plugin-help"
-    ]
+    ],
+    "repositoryPrefix": "<%- repo %>/blob/<%- version %>/<%- commandPath %>"
   },
   "main": "src/index.js",
   "scripts": {
-    "posttest": "eslint src test e2e",
+    "pretest": "eslint src test e2e",
     "test": "npm run unit-tests",
     "unit-tests": "jest --ci",
     "prepack": "oclif-dev manifest && oclif-dev readme",
