+++ test/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "rules": {
+      "node/no-unpublished-require": 0
+  }
+}
